fix cz80 hi memory bug (BasePC < 0); fix crash caused by memory header; now most (CPS1) games run
git-svn-id: https://blade.imwork.net/svn/fba4dstwo@9 40d471ee-5a4d-f447-990a-021eff10c44c

+++ src/cpu/cz80/cz80.h
@@ -90,7 +90,7 @@ extern "C" {
 #define zlSP		CPU->SP.B.L
 #define zhSP		CPU->SP.B.H
 
-#define zRealPC		(PC - CPU->BasePC)
+#define zRealPC		(PC - (unsigned int)CPU->BasePC)
 #define zPC			PC
 
 #define zI			CPU->I
